Moving response_headers.py: browse.services.util -> browse.controllers
