+++ tests/test_304.py
@@ -5,7 +5,7 @@ from datetime import timedelta
 from dateutil import parser
 
 
-from browse.services.util.response_headers import mime_header_date
+from browse.controllers.response_headers import mime_header_date
 
 from app import app
